context-management/pyproject.toml
[project]
name = "context-management"
version = "0.1.0"
description = "Intelligent Context Management System for BigAGI"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "tiktoken>=0.8.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
